refactor: move package internal.player to bukkit.player
